+++ crates/ui/src/lib.rs
@@ -1,10 +1,12 @@
 mod components;
+mod element_ext;
 mod elements;
 mod modules;
 pub mod prelude;
 mod theme;
 
 pub use components::*;
+pub use element_ext::*;
 pub use elements::*;
 pub use modules::*;
 pub use theme::*;
